rproxy,mediawiki,leaps: Move tests next to module
We now allow defining paths when listing tests in a module, so let's
move the reverse-proxy, mediawiki and leaps test to their actual module
implementation.

What's still a bit ugly is the import of eatmydata.nix, but this could
be cleaned up by creating a proper NixOS module for that.

Signed-off-by: aszlig <aszlig@nix.build>