modules/web: Rename backends to webserver
I'm renaming this to "webserver" in singular form because we already
have "database", which is singular as well. This also copes well with
the option namespace which was "webserver" already.

The "backend" option is now called webserver.variant, so we don't clash
with the "webserver" namespace. So changing the backend is just a matter
of eg.: webserver.variant = "lighttpd";

Another change for this rename is that the backendInit option is now
called webserver.init and startupScript now also resides under the
"webserver" namespace.

Signed-off-by: aszlig <aszlig@nix.build>