webserver: Fix unit for webserver-init
If the config option webserver.init is given but no startupScript the
unit gets an empty script argument, which in turn causes systemd to not
start the unit (and thus the preStart script is omitted).

This particularly led to a failure of the custom-webservice test, which
only defines webserver.init (in lighttpd.nix) and no startupScript.

Signed-off-by: aszlig <aszlig@nix.build>