mkUnique: Don't add suffix if suffix equals wsName
For example if you have a web service for leaps which is called "foo",
the user and group name for the leaps service would end up being
"leaps-foo-leaps", which is a bit redundant.

So with this change, the user name is now just "leaps-foo" instead.

Tested via the leaps webservices test.

Signed-off-by: aszlig <aszlig@nix.build>