modules/apache: Make allDenied/allGranted DRY
Having version checks for this in all of the submodules really doesn't
make sense if we have a central point (the webserver module) where we
can define a module argument for Apache-related library functions.

This is done now, and we have an "apache" module arg available for all
webservice submodules.

In the long run however, we might want to abstract away the various
webserver implementations so we can declaratively specify the paths we
want to have access and the various locations to configure.

In the end the idea would be to switch webserver implementations using a
single line in the configuration.

Signed-off-by: aszlig <aszlig@nix.build>