modules/web: Implement assertions for meta.license
Every single webservice module needs to have a meta.license attribute
set, otherwise NixOS evaluation will fail now with an assertion.

The reason this works on a per-webservice basis is because we have a
submodule, so we ensure that there is only *one* option defined within
that submodule and we also check that by passing an assertion up to the
top-level configuration.

Signed-off-by: aszlig <aszlig@nix.build>