release.nix: Add a job for the options manual
This manual only contains all the nixcloud-related options and I'm
adding this mainly so that the description for our NixOS options are
validated correctly.

Apart from that this also has the advantage that it's easier to discover
which options are actually there without having to resort to evalution
and/or rely on some external service.

Signed-off-by: aszlig <aszlig@nix.build>