Add a module for creating dirs and setting perms
So the first step is doing this at the top-level, so that we can pass
through a subset of the per-instance options (which will be introduced
later) to our mechanism.

The mechanism in itself is based on a directories option which consists
of an attribute set of paths and their corresponding owner/group and
additional permissions to set during creation and also during fixup.

While the fixup phase is run whenever the directory already exists the
creation phase is run whenever the directory does not exist.

Right now the implementation is pretty much work in progress, but I'm
already committing this straight to master because whenever the
directories option is not used the module doesn't do anything.

Also, we have several limitations left to solve before this is
production-ready:

 * The creation and fixup units need to have a conflict on each other so
   that we don't run into a race condition between those two.
 * Currently the permissions.recursive option doesn't do anything.
 * More tests need to be implement to make sure we properly propagate
   default ACLs.
 * We need to fix up the generated systemd services so that they get
   properly ordered and don't start in parallel with the services that
   require them.
 * Make sure we don't open up ways for privilege escalation. For example
   if the fixup service first sets the permissions and the owner/group
   later, this could lead to elevated permissions of the previous
   owner/group and/or the allowed users/groups in the ACL.

In order to implement the permissions.recursive option, we need to take
care of nested directories, so let's say we have a directory /a/b and
/a/b/c, we need to stop recursing during fixup of /a/b once we hit
/a/b/c so that we don't override the permissions from that subdirectory.

Signed-off-by: aszlig <aszlig@nix.build>
Issue: #6