directories: Bring defaults in par with its docs
We actually have documented sensible defaults, but the actual
permissions that are set do not reflect the documentation.

Alongside of that, the documentation itself does have some
inconsistencies and strange wording, so this also fixes them.

With this change, we now pass an attribute set to mkPermOpts, so we
can't accidentally set the wrong flag without noticing because we now
need to explicitly specify something like eg. "write = false" instead of
counting the argument position.

Signed-off-by: aszlig <aszlig@nix.build>