directories: Don't create dirs with mode 0000
Using mode 0000 for the last component of the directory would be fine
because we're going to grant the permissions explicitly afterwards.

Unfortunately if one of the parent directories does not exist it will
also end up having mode 0000 and additional permissions are *not*
granted afterwards.

So this ends up with users having no access to the directory at all,
except of course the root user.

We now use mode 0755 by default and the leading dirs then are also owned
by user root and group root, so there is no way a normal user to
escalate privileges between creation and granting permissions.

Signed-off-by: aszlig <aszlig@nix.build>