directories: Implement recursive fixup of perms
This is one of the main features of this module and we really need to
take care to not overwrite permissions of subdirectories that are also
defined via the same module option.

In order to avoid this, we filter all the subdirectories from the
directory definitions and exclude them using "find" arguments (-path and
-prune).

The rest of the find arguments is to ensure that directories and files
are treated differently, because files usually shouldn't be executable.

I've also added a few tests to ensure that recursive fixup of
permissions is working correctly and whether changing of the owner/group
works as well.

Signed-off-by: aszlig <aszlig@nix.build>
Issue: #6