directories: Allow to configure default dir mode
So far parent directories were always created with mode 0755, but we
might want to configure this. The main reason why the parent directories
had 0755 is not because we used -m to mkdir but because the default
umask is 0022, which in turn leads to mode 0755 for directories.

Using the -m flag for mkdir only applies the mode to the final path
component.

Our newly introduced option defaultDirectoryMode now allows to set an
arbitrary mode, so we now have a small shell script which is creating
parent directories one-by-one instead of using -p so that we can apply
the correct mode. Another possibility would have been to calculate the
right umask for the mode given, but this won't work if the mode has
setuid, setgid or sticky flags.

Right now the test fails randomly because we still have to fix ordering
between mkdir units.

Signed-off-by: aszlig <aszlig@nix.build>