directories: Set permissions for owner/group
So far we did actually set the owner and group permissions, but only as
a direct ACL entry by listing its name. The latter however isn't
sufficient because if one of the user/group acl entries apply to the
same user/group as the owners of the file, the owners will have
precedence.

Signed-off-by: aszlig <aszlig@nix.build>