directories: Fix ordering of mkdir units
We want to create directories beginning from the root directory all the
way down to all the leaves. Creating subdirectories before the parent
directories or even parallelize that will lead to random failures, so we
have to ensure the right order.

This also fixes the VM test so that it's now stable and doesn't fail at
random.

Signed-off-by: aszlig <aszlig@nix.build>