modules/web: Integrate directories module
While we already have splitted up the directories module to be directly
included inside the submodule in order to re-use it, it didn't turn out
to be the best idea.

Instead, we now only use the options but not the actual implementation
along with the services. We now pass the webservice directory
definitions to the top-level by also prepending the stateDir and
removing the instance.* options.

This allows us to have one module where we have *all* of the mkdir/fixup
units in one place so that it's easier to find and avoid collisions.

Signed-off-by: aszlig <aszlig@nix.build>