directories: Avoid race conditions during mkdir
We do a check whether a directory component exists and if it doesn't we
create it, which unfortunately upens up a race condition between the
check and the creation.

This usually shouldn't be a problem if *only* the directories module is
in charge of creating directories, which is what's the goal.

Unfortunately this also has a second problem, mainly that whenever we
have directories with a common ancestor we run into the same problem.

For example if we have /common/foo and /common/bar, the /common
directory will be subject to the same problem.

In the end we really should handle common ancestors, but using
"mkdir -p" even for single components is safer as well because it's an
atomic operation. My initial fear that it might create more than one
component is also not really the case because that recursive function
already steps through components.

So the only occasion where this could still run into a race condition is
when a directory is removed while one of the mkdir services is creating
components.

The impact of the last occasion would be that the permissions given in
defaultDirectoryMode simply aren't applied for one or more components of
the path.

I've added a small check within the directories VM test, which is by no
means exhaustive and only checks for hard failures to create
directories, but it doesn't yet make sure that permissions of leading
directories are correct.

Spoiler warning here: If we were to check for that, the test would fail
randomly, because our implementation currently doesn't handle common
ancestors.

Signed-off-by: aszlig <aszlig@nix.build>