modules/web: Switch to using directories module
This finally uses the new directories implementation, which means that
we now can finally inherit permissions in a proper way (rather than just
using the setgit bit).

It also gets rid of a few webserver init scripts which are no longer
necessary.

I've tested this against all of the VM tests.

Signed-off-by: aszlig <aszlig@nix.build>
Fixes: #6