modules/web: Convert createHome to directories
This of course only applies if the user's home directory is within
stateDir, but whenever this is the case a directory definition is
created for it and createHome is passed as false to the top-level
users.users option.

Passing the users home directory to the directories option also has the
side effect that it overrides existing directory options, so in order to
prevent this, we pass those values with a priority of 200 so that other
defined options by the user will have precedence.

I've also added a test which ensures the ownership of the stateDir in
the leaps web service.

Signed-off-by: aszlig <aszlig@nix.build>
Issue: #6