Add HSTS and CSP to example redirect
Redirects from http are inherently not very safe (there is a 'trust gap' as the plain text request redirects). If a user has some http url hard-coded in say some application, the unsafe link is used again and again. Over time she or he might run the risk of opening that link in an unsafe environment. By using HSTS, the browser retains the knowledge that it was rerouted to https in the safe environment - and that it should always open the https version. Since the redirect does not need any content elements, CSP can be used to forbid any content elements to end up in the browser.