web/services/README: Move to documentation/
My initial thought would be that developers would find this README in
the services directory and it's displayed directly when browsing that
directory on GitHub.

However, I think most users would probably look into documentation to
find these things and not browse the services directory on GitHub.

If we still want it to show up in services/, we could still do a symlink
I guess (not sure whether GitHub would display it correctly).

Signed-off-by: aszlig <aszlig@nix.build>
Issue: #11