Add an escape function for DNS labels
I'm implementing the DNS functionality from bottom-up, so let's start
with labels. These labels can contain arbitrary 8-byte characters and
most of them are escaped via decimal escapes.

See https://tools.ietf.org/html/rfc4343#section-2.1 for the escaping
rules of labels.

Implementing such an escaper in Nix itself using stringToCharacters and
concatStrings would be quite inefficient and we're going to have lots of
labels. So instead, we now have a generator which emits a Nix expression
file containing a table of match values as a list and a table of replace
value as a list, currying builtins.replaceStrings.

Signed-off-by: aszlig <aszlig@nix.build>