static-{darkhttpd,nginx}: Clean up tests a bit
We don't need libeatmydata, because we're not using a DBMS. Furthermore,
let's use a plain comparison instead of grep, because it's much more
precise, as grep is searching for a string at any point within the text.

Also, we don't need quoting on "works", because it doesn't contain any
special shell characters and it also doesn't make much sense to use a
special Nix attribute for that.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight