web: Hash user/group names longer than 31 chars
From the manpage of useradd(8), the limit described there is 32
characters, but it turns out that systemd's valid_user_group_name()
limits this to UT_NAMESIZE - 1, which is 31 characters.

So we now have two more functions available for multiple instance
modules, mkUniqueUser and mkUniqueGroup, which both do the same at the
moment. The reason why I made separate functions is that we then have it
easier in the future if we want to handle group names differently.

Getting back to the actual implementation, we check whether the
user/group name is more than 31 characters long, and if it's the case we
create a SHA256 hash, truncate it to 26 characters and add "user-" in
front of it. The reason why we do that is that user names may not begin
with a number while hashes perfectly can.

I've also added a test which checks whether a service with such a user
name is still starting up.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight
Fixes: #15