database: Add a database shell helper utility
In order to connect to a database that's residing within a web service
so far a lot of knowledge about internals was necessary.

So this now makes it way easier to connect to a database, by adding a
helper utility which not only connects to the database itself but also
is able to figure out the database type.

The latter is done by providing an internal option to the database
module which specifies the command necessary to connect to it.

All the mapping is then done using a new "dbshell" module, which
generates the utility from the configurations of all web services and
exposes it to the system path.

A command like "nixcloud-dbshell mediawiki-foo mediawiki" will now
connect to the "mediawiki" database within the web service with the
"mediawiki-foo" unique name.

Signed-off-by: aszlig <aszlig@nix.build>