docs: Describe usage of database module
This fixes a few FIXME-placeholders that were in the documentation.

While I'm not sure whether I addressed "give usage example from
./modules/web/services/mediawiki/default.nix on how to access either"
correctly, but the maintenance process should be documented sufficently.

I kept the last placeholder about writing documentation on how to create
database modules intact, because it's currently everything but easy to
do and relies heavily on the internals of the main module. This is why,
for now, I decided to not document the whole process here, because the
internals are documented in the source code and its implementation might
change drastically once we have a reliable way to generalize the
services for the database creation/population.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight