dbshell: Factor out mapWebServiceConfigToList
This is also used by the root web service module and might come in handy
for other modules, so let's actually provide it as a module argument in
top level modules.

The reason why I'm not also using this function while calling
make-toplevel-config.nix is because the latter would run into an
infinite recursion on nixcloud.webservices. So I've de-generalized that
part into the "toplevel" attribute.

Signed-off-by: aszlig <aszlig@nix.build>