dns/escape-label: Don't hardcode replaceStrings
We need to escape and unescape labels based on whether we want to
reference a particular node or whether we want to generate a zone file.

By allowing to pass a function as an argument to this expression file,
we're able to use either "replaceStrings" or "flip replaceStrings" to
escape and unescape respectively.

Signed-off-by: aszlig <aszlig@nix.build>