dns: Implement core types and options
This adds the base module which is responsible for generating the tree
of option declarations and definitions that we need in order to
eventually generate zone files.

The idea here is to let the user specify labels in an intuitive way and
we're generating a tree on that basis using types.coercedTo.

So in the end if the user supplies something like "a.a.a", the internal
tree generated will be "subZones.a.subZones.a.subZones.a".

Right now the record types are more or less a placeholder I used for
testing, so we need to overhaul these. We also need to bring in our
previously written functionality for handling labels.

Signed-off-by: aszlig <aszlig@nix.build>