dns: Implement type for domain
For checking values this type is pretty much straightforward and it
checks whether the splitted labels have the right size and the count of
the labels is correct.

However, for merging we do have a distinction between absolute and
relative domains, which we can use to look them up in our DNS hierarchy
tree and also use it to escape values during zone file generation.

I've also rewritten splitDomain, because using builtins.split doesn't
allow us to do assertions, so if we want to match something like ".."
properly we either need to go a even more complicated route or simply
convert each character to a list and work on that. The latter is still
not very efficient but we should have low usage for that function.

Signed-off-by: aszlig <aszlig@nix.build>