dns/lib: Add type for IPv4 and IPv6 addresses
This might also be useful outside of the DNS implementation, but for now
let's use this for validating A and AAAA entries.

The actual implementation is quite messy, because parsing IPv6 addresses
isn't easily done using regular expressions. It's not that it's
impossible to do but doing it using a regular expression is even messier
as it boils down to a single big regex that uses lots of alternation.

So the main reason for actually validating those addresses as opposed to
just using types.str is that we want to avoid generating invalid zone
files at all cost.

Signed-off-by: aszlig <aszlig@nix.build>