dns: Implement common record types
With all the digging through DNS RFCs I probably have lost track about
what are *really* the most common resource record types, so hopefully
nobody is killing me if I forgot about one.

Anyways, the way these record types are represented are by using a
submodule underneath, which also exposes the raw resource record entry
to the top level.

Right now the way those are structured are in a non-uniform way, so in
the next step we either need to either make them uniform on the
individual option level (which I think will be hard to do) or use a
function which takes care of that from the base module.

Another thing that's particularly icky is that the CAA record currently
is a bit of an exception and doesn't fit well with the generalization
written to represent resource records.

I've also updated the example to reflect those changes, however as a few
things are still incomplete we might need to alter the example again
very soon.

Signed-off-by: aszlig <aszlig@nix.build>