dns: Fix types for CAA records
This is one of the resource record types which deviates a bit from the
usual way resource records are defined in DNS, so specifying the types
for them is a bit more involved as we have to generate the property
tags.

So this should fix evaluation of CAA records by actually implementing
its options.

Signed-off-by: aszlig <aszlig@nix.build>