dns: Accumulate resource records in records module
This is so that we have a list of resource records available from within
the records submodule, which helps to turn the hierarchy into a flat
list structure that we can use to generate individual zone files (or an
equivalent form in name server implementations other than bind or nsd).

We still need to write the functionality to actually traverse the tree,
but at least we no longer need to handle internal details for the
records module while doing so.

Signed-off-by: aszlig <aszlig@nix.build>