dns: Generate a flat list of defined zones
With this, we should have every information that we need to feed to the
zone file generator that's about to be implemented soon, so the zoneList
attribute isn't used yet.

Here I also added another isZone option, because it's difficult to
reliably determine whether a specific option is defined when the actual
option is within a submodule but we want to get that information from
outside of the submodule.

Signed-off-by: aszlig <aszlig@nix.build>