dns/lib: Add functions for joining domains
Internally we have passed through all the domains specified by the user
using lists of labels, but in order to generate a zone file we need to
get them back into a single string.

So we now have joinDomainSimple, which simply joins them and escapes the
labels and the joinDomain function, which does the same but also handles
domain.absolute and domain.relative values.

Signed-off-by: aszlig <aszlig@nix.build>