dns: Correctly encode list of scalar records
Multiple resource records for non-submodule types such as A or AAAA can
be specified like this:

nixcloud.dns.zones.org.example.A = [ "1.2.3.4" "5.6.7.8" ];

This should result in two records being created, however in our
implementation it didn't but created two RDATA elements.

Of course, this is not what we want and will lead to an invalid zone
file, so let's use a scheme that is similar to mkRecordModule allowing
to pass a singleton flag if the resource record is something that can
only occur once (like for example CNAME).

Signed-off-by: aszlig <aszlig@nix.build>