dns: Fix assertion on CNAME
A CNAME record can't have other resource records for the same domain, so
for checking this I gathered an attribute set from the record type name
to booleans which indicate whether the option is defined.

However while checking, I accidentally mapped over all of the attribute
names, so whenever you defined a CNAME record in a valid way the
assertion would trigger with an error listing all the other resource
record types even if they're not defined.

Signed-off-by: aszlig <aszlig@nix.build>