dns: Correctly detect whether zones are defined
The outermost type of nixcloud.dns.zones is a submodule, so comparing
the configuration value with an empty attribute set doesn't work,
because the value always contains the submodule options.

So instead, we remove the default and use options.foo.isDefined instead,
which for submodules works much better than what we did so far.

Signed-off-by: aszlig <aszlig@nix.build>