dns: Add integration test
We use nixcloud.dns.zoneList to build up a series of dig commands that
check whether the records specified in the example config (example.nix)
are matching up the values reported by dig.

The dig tool is part of the BIND tools, so the syntax doesn't differ
very much, but if we have a DNS server implementation that doesn't use
BIND master file syntax the test becomes even more valuable.

Signed-off-by: aszlig <aszlig@nix.build>