Implement new version module
First of all, the module adds the current revision of nixcloud-
webservices to nixos-version and in addition to that, there is now a
small helper utility called "nixcloud-version" which just prints the
current git revision of the nixcloud-webservices version in use by the
system.

If the module is unable to retrieve the git revision, "master" is
suffixed instead so that we have a reasonable default rather than just
failing.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight