database: Prepend "localhost:" to socket path
Some database abstractions try to be very "clever" by looking at the
host name to see whether it is an IPv6 address instead of just passing
it verbatim to PHP. This for example is the case with Nextcloud.

So in order to address this, let's prepend localhost to the actual path,
so that it won't be treated as an IPv6 address.

Signed-off-by: aszlig <aszlig@nix.build>