database: Implement setup hooks for temporary DBs
This is very useful for preinitializing database schemas, so that we can
avoid as much errors as possible during runtime and fail at build time
instead.

The implementation currently is not yet stable and might change, so I
marked most of the options as internal for now.

Using those setup hooks however is already possible by just adding
config.database.myshiny.tempDbSetupHook to nativeBuildInputs of a
derivation.

After unpacking, the database is started up and the tempdbShell and
tempdbDump functions can be used afterwards. The former is for getting a
database shell to the temporary instance and the latter dumps the
database schema.

Tests for these hooks are currently missing and I'll add them once the
interface is stable and documented.

No other functionality should be affected by this change and I've
verified that it doesn't introduce regressions by running all of the
tests.

Signed-off-by: aszlig <aszlig@nix.build>