lib/call-test: Allow to define unit tests
So far we are only able to define VM tests, but for testing individual
functionality it really would make sense if we'd be able to run unit
tests in a Nix instance within the build.

However, implementing this in a generic way is somewhat tedious, because
within the build we don't have access to arbitrary paths and also need
to replace paths to their relative counterparts, which itself is a bit
error-prone.

Fortunately, for basic unit testing this should work well, because we
stay entirely within nixcloud-webservices and nixpkgs.

In order to distinguish unit tests from VM tests, we now have a "type"
attribute which either is absent (then it's a VM test) or it's "unit".

Signed-off-by: aszlig <aszlig@nix.build>