lib/call-tests: Prepend test_ to unit test attrs
It really doesn't make sense to collect only those attributes which have
a name prefixed with "test", because we have let and recursive
attributes so we can simply ensure that only real unit tests enter the
main attribute set.

So what we're doing now is prepending "test_" to all attribute names, so
that the actual test runner picks up all of them.

Signed-off-by: aszlig <aszlig@nix.build>