web/core: Allow nesting of mkUnique/mkUniqueUser
We have a few occasions where we use mkUniqueUser or mkUniqueGroup which
are then fed through mkUnique when generating systemd units for example.

In order to prevent the user name length to exceed the maximum length we
need to ensure that mkUnique doesn't alter the hashed values in any way.

Signed-off-by: aszlig <aszlig@nix.build>