Add unit tests for mkUnique*
So far we had the username-hashing test, which was a VM test and only
checked *one* value. However, with unit tests in place we can make those
tests much more exhaustive.

For example we now generate tests (in a similar vein to Haskell's
SmallCheck) that generate user name strings up to 200 characters and
checking whether the results are correct and/or whether a nested call of
mkUnique, mkUniqueUser, mkUniqueGroup alters these values.

Signed-off-by: aszlig <aszlig@nix.build>