release.nix: Allow to work in restricted eval mode
Hydra in conjuction with Nix 2.0 and restricted eval mode mode now
doesn't allow to use paths from NIX_PATH anymore but everything has to
be propagated down from release.nix.

So this fixes a few of the most obvious occurences, but there might be
more things we need to fix up after the next Hydra evaluation.

Signed-off-by: aszlig <aszlig@nix.build>