Fix database startup ordering
This basically just takes out multi-user.target from the wantedBy list
of database.target. The reason is that whenever a service has databases
enabled but doesn't pull in database.target we will run into a problem
because creation of databases will be done without actually starting up
and/or initialising the DBMS.

I've also added a small comment to illustrate the startup sequence to
make it more clear to people wanting to implement their own DBMS
modules.

Signed-off-by: aszlig <aszlig@nix.build>