directories: Add postCreate and postCreateAsRoot
These options allow to link a command to execute to creating a
directory, which is useful for initializing other files that should be
inside that directory.

An example for this would be to create an initial index.html in a web
root.

Both options do the same but postCreate runs it at the permissions of
the directory owner and postCreateAsRoot runs it with the permissions of
the superuser.

I tried to first integrate this into the mkdir services, but setting
user and group attributes within one unit is a bit wonky to do when you
don't want to run with PAM. So I separated the individial postCreate
units into a target with the negative condition of the existence of the
corresponding directory and it didn't work out either.

The reason is that units depending on another unit with a condition will
be started anyway, which is by design, see:

https://github.com/systemd/systemd/issues/985

A way to implement this would be to use assertions, which then cause the
triggering unit to fail. Unfortunately this will loudly log an error and
thus is confusing for users.

In the end I ended up with a postStart script which invokes systemctl to
start up the postCreate units in sequence.

Signed-off-by: aszlig <aszlig@nix.build>