Add a small run-webservice helper
This is based of my original implementation at [1], but with a bit of
cleanup and comments. The cleanups revolve mostly about
depersonalization, because the original implementation contained my Zsh
and Vim configurations.

Also, the original implementation doesn't work very well on non-NixOS
systems because the permissions for the Nix store are not necessarily
the same. So we patch OpenSSH here to not do the permission check on the
private key.

Another deviation from the original implementation is that we are now
more thoroughly checking whether the SSH daemon is up and running rather
than just checking whether the TCP port is available. The reason the
simple TCP port check doesn't work always is that during early
initialization of OpenSSH the connection is dropped and that in turn
causes the VM to shut down.

Unfortunately I had to do a few workarounds to ensure the reverse proxy
sets the correct hostname, so the actual webservice gets the same port
as the forwarded port on the host of the VM, while the reverse proxy
itself binds to port 80.

[1]: https://github.com/headcounter/habitica-standalone/blob/bbe5a25529c7957ecc2974915ed84118faa0067e/build-vm.nix

Signed-off-by: aszlig <aszlig@nix.build>