static-darkhttpd: Add an example index.html
This is to direct users which didn't configure anything on what to do
next. I also changed the default root to "${stateDir}/www", because
stateDir itself may contain other files that you most likely do not want
to expose.

If the default root is set to something else than the default, no www
directory is created within stateDir.

Signed-off-by: aszlig <aszlig@nix.build>
Issue: #14