testing: Use pkgs.path instead of <nixpkgs>
We're passing through the nixpkgs source from Hydra in order to
reference other files within that source tree, so for referencing files
there is the nixpkgs argument.

Unfortunately if we don't have the path referenced by <nixpkgs> in the
store, evaluation will fail, so let's actually turn the pkgs argument
into an explicit import of <nixpkgs> by default (so it doesn't
necessarily need to be in the store) and let Hydra override this with
the real store path. That way it should work with both restrict-eval and
nonrestricted eval.

Signed-off-by: aszlig <aszlig@nix.build>