email: Turn main submodule into normal options
The added evaluation overhead really isn't needed here, plus it makes
the code more unreadable and can lead to confusion when checking whether
an option is defined.

For example if you want to check whether nixcloud.email.ipAddress is
defined you can only do that from within the submodule but not from the
top level (except of course if you pass through options).

Signed-off-by: aszlig <aszlig@nix.build>