email: Clean up main module expression a bit
This indents the descriptions a bit as they're hard to read when they go
too far to the right. In addition I've also fixed a few them, because they
weren't valid Docbook (especially when it comes to cross-referencing
options).

More importantly however, I added types to all the options because I
already stumbled on types being merged in the wrong way while writing
the tests. One example would be "ipAddress" which would concatenate the
strings if it's defined more than once instead of producing an error.

I also reordered the long mkMerge block to have all the conditions at
the top and all stuff without conditions at the bottom so it's less
evaluation overhead and also easier to find stuff. During reordering I
also found some wrong indentation which I fixed as well.

Another thing that I have found is that "incoming" was spelled
incorrectly ("incomming") and both the wrong spelling and the right one
was spread throughout the code.

Signed-off-by: aszlig <aszlig@nix.build>