email: Add a very basic VM test
This is a bit more involved than I initially thought and I first started
to get the test runtime down to something where you don't die waiting
for timeouts. The main bottleneck here are DNS timeouts, so the test
network now contains a name server that serves the root zone for all the
other nodes. I wanted to use NSD for this, but being able to serve the
root zone in NSD is a compile-time flag, so I used BIND for that because
we don't want to shove down compiling NSD to the users throats every
time they update their system or initially enable nixcloud.email.

However, it seems to me that the BIND NixOS module isn't used by anybody
for anything more serious than just configuring a few zones or use a bit
of the options that are provided by the module, because we needed to use
a VERY ugly workaround just to inject options within the "options"
section of named.conf.

My initial implementation was comprised of a combination of Swaks[1] and
cURL to send and fetch emails, but it became very tedious to handle even
the simplest stuff. The reason why this falls short is because it relies
very heavily on the shell and we need wait for new mails of the user's
inbox (possibly in the background). While Swaks is quite comprehensive
the IMAP support of cURL is very bare-bones, so the result would have
been a mixture of Perl, Shell and Nix with lots of nested shell quoting
and other nuisances.

So I decided to implement the tests in Python because it already has
extensive support for SMTP, IMAP, POP3 and MIME in its standard library.

The resulting test suite still has its rough edges as we have a lot of
redundancy, but at least we now have something to build upon. In an
ideal world we'd have a testing framework similar to Escalus, which
makes story based tests quite compact and readable but I guess I'm tired
and starting to dream.

For now I have disabled greylisting during the tests, because waiting
for >300 seconds until the test completes is going to get everybody into
a serious depression along with self-harm and possibly suicide. That's
obviously something we do not want :-)

[1]: http://www.jetmore.org/john/code/swaks/
[2]: https://github.com/esl/escalus

Signed-off-by: aszlig <aszlig@nix.build>