email: Make it actually possible to disable TLS
So far, setting enableTLS to false just gets rid of the certificates and
leaves Postfix in some kind of limbo where it tries to enforce TLS but
if the client wants to use TLS it also fails because the certificates
aren't there.

This is now no longer the case and the reason why we need to disable
this in the first place is that we need it for our first version of the
test, at least until we have everything set up so that we can
successfully test TLS as well.

Signed-off-by: aszlig <aszlig@nix.build>