TLS: Clean up module expression a bit
The file is very hard to read, especially because of the indentation
mismatches in braces and parens.

I've also fixed all the descriptions to contain proper Docbook, as a few
of them were using characters that are invalid in XML. So this should at
least fix building the manual for "nixcloud.TLS".

The tls_certificateSetModule attribute is now also gone as it's not used
anywhere in this file.

I've made no functional changes other than fixing the manual for *this*
module (there still is another error but that one is in another module).

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight