reverse-proxy: Fix undefined reference to 'domain'
Using isDefined here clearly is just a workaround for now, but it fixes
the evaluation error if the option is not defined. In the long term
however, it makes more sense to actually either set the default value of
domain or not set a default for TLS altogether.

I've also used "toplevel" here, even though it's not really the toplevel
and the attribute is completely unnecessary, but that's for another
cleanup someday.

I did however fix the description of the TLS option while being at it.

With this commit, the manual should build successfully again.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight