database: Re-add dependency on multi-user.target
Regression introduced by me in 93d1cda690f289169a792cf4aa41991274f132d1.

In theory this would be a good idea to leave out multi-user.target but
we don't have any services that pull in database.target directly and
they also shouldn't need to because that's why we can enable or disable
databases.

So I'm adding back the dependency for now until I have found a good
reason not to do so. Apparently I must have had one then but I didn't
note it in the commit message, which is quite an epic failure (given
that I usually write quite detailled commit messages) to say the least.

Signed-off-by: aszlig <aszlig@nix.build>